close to finish class implementation

+++ src/main/java/org/alshar/lib/data_structure/CoarseningEdge.java
@@ -1,5 +1,5 @@
 package org.alshar.lib.data_structure;
 
 public class CoarseningEdge {
-    float rating;
+    double rating;
 }
